Move utils.deprecated.is_v1() and .generate_otu_dirname() to legacy.utils
