+++ virtool_cli/utils/deprecated/reference.py
@@ -1,19 +1,5 @@
-import re
 from pathlib import Path
 from typing import Optional, Tuple
-
-
-def is_v1(src_path: Path) -> bool:
-    """
-    Returns True if the given reference directory is formatted under version 1 guidelines.
-    Determines if virtool ref migrate should be run before using this version of virtool-cli
-    
-    :param src_path: Path to a src database directory
-    :return: Boolean value depending on whether alphabetized bins are found.
-    """
-    alpha_bins = list(src_path.glob('[a-z]'))
-
-    return bool(alpha_bins)
 
 
 def get_otu_paths(src_path: Path) -> list:
@@ -43,9 +29,7 @@ def get_sequence_paths(isolate_path: Path) -> list:
     :param isolate_path: Path to an isolate directory under an OTU directory
     :return: A list of paths to all sequence files in an isolate directory
     """
-    sequence_ids = [
-        i for i in isolate_path.glob('*.json') if i.stem != "isolate"
-    ]
+    sequence_ids = [i for i in isolate_path.glob("*.json") if i.stem != "isolate"]
 
     return sequence_ids
 
@@ -58,31 +42,11 @@ def search_otu_by_id(otu_id: str, src_path: Path) -> Optional[Path]:
     :param src_path: Path to a reference database directory
     :param otu_id: Virtool OTU unique ID
     """
-    for path in src_path.glob(f'*--{otu_id}'):
+    for path in src_path.glob(f"*--{otu_id}"):
         if path.is_dir():
             return path
-    
+
     return None
-
-
-def generate_otu_dirname(name: str, otu_id: str = '') -> str:
-    """
-    Takes in a human-readable string, replaces whitespace and symbols
-    and adds the Virtool hash id as a suffix
-    
-    :param name: Human-readable, searchable name of the OTU
-    :param otu_id: ID hash of OTU 
-    :return: A directory name in the form of 'converted_otu_name--taxid'
-    """
-    no_plus = name.replace('+', 'plus ')
-    no_symbols = re.split(r'[():/-]+', no_plus)
-    joined = ' '.join(no_symbols)
-    no_whitespace = re.sub(r'[\s]+', "_", joined)
-
-    dirname = no_whitespace.lower()
-    dirname += '--' + otu_id
-
-    return dirname
 
 
 async def get_unique_ids(otu_paths: list) -> Tuple[set, set]:
